Update Anton and Digits.cpp

+++ 800/Anton and Digits.cpp
@@ -7,7 +7,7 @@ using namespace std;
 	{
 		ll a,b,c,d;	
 		cin >> a >> b >> c >> d;
-		int kp = min(a, min(c, d));
+		ll kp = min(a, min(c, d));
 		cout << kp * 256 + min(a - kp, b) * 32;
 		return 0;
     }
